feat: 컬러 에셋이름 변경 & 포우컷 선택 뷰 UI 구현
