+++ PawCut/PawCut/Sources/DesignSystem/Font/FontSet.swift
@@ -25,6 +25,7 @@ enum FontSet {
         case _14 = 14
         case _16 = 16
         case _17 = 17
+        case _18 = 18
         case _20 = 20
         case _24 = 24
     }
